Moved most of move() to array ops.
basically just postponed the problem of splitting it up. I realized alot of the hot ops need the array contexts, right now what we're doing is passing them outside of the organism class which I really hate. It doesn't really count. It helps with readability for now.

ToDo: actually split it up